Add addBook method
