+++ modules/ui.js
@@ -1,4 +1,4 @@
-export class UserInterface {
+export default class UserInterface {
 
   static displayBooks(){
     const storedBooks = [
@@ -17,5 +17,36 @@ export class UserInterface {
     books.forEach(book => UserInterface.addBook(book));
   }
 
+  static addBook(book) {
+    const bookList = document.querySelector('.book-list');
+    const bookDiv = document.createElement('div');
+    bookDiv.classList.add('book-holder');
+    // add book title
+    const titleP = document.createElement('p');
+    titleP.textContent = `"${book.title}" by \xa0`;
+    // add author
+    const authorP = document.createElement('p');
+    authorP.textContent = ` ${book.author}`;
+
+    const paraHolder = document.createElement('div');
+    paraHolder.appendChild(titleP);
+    paraHolder.appendChild(authorP);
+    paraHolder.classList.add('group-paragraph');
+
+    const bookId = document.createElement('p');
+    bookId.setAttribute('type', 'hidden');
+    bookId.setAttribute('value', book.id);
+    //  create delete button
+    const deleteBtn = document.createElement('button');
+    deleteBtn.textContent = 'Remove';
+    deleteBtn.classList.add('deleteBtn');
+    // const line = document.createElement('hr');
+    bookDiv.appendChild(paraHolder);
+    bookDiv.appendChild(bookId);
+    bookDiv.appendChild(deleteBtn);
+    bookList.appendChild(bookDiv);
+
+    bookList.appendChild(bookDiv);
+  }
 
 }
